skip: remove skipstorylevel and add remaining modules
